pyproject.toml
[project]
name = "genrepo"
version = "0.1.5"
description = "Add your description here"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "License :: OSI Approved :: MIT License",
]
authors = [
    { name = "Andrea Fuentes", email = "mfuentescastellanos@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "typer>=0.21.1",
]

[project.optional-dependencies]
sqlmodel = ["sqlmodel>=0.0.32"]
sqlalchemy = ["sqlalchemy>=2.0"]

[project.scripts]
genrepo = "genrepo.cli.app:main"

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "mypy>=1.11.0",
    "build>=1.4.0",
    "twine>=6.2.0",
]

[tool.ruff]
exclude = [
  "tests/golden",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy"]
exclude = [
  "^tests/golden/",
]

[[tool.mypy.overrides]]
module = [
  "typer.*",
  "rich.*",
  "jinja2.*",
  "sqlmodel.*",
  "sqlalchemy.*",
  "yaml",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "tests.golden.*",
]
ignore_errors = true

[tool.uv]
index = [
  { name = "pypi", url = "https://pypi.org/simple" },
  { name = "testpypi", url = "https://test.pypi.org/simple", publish-url = "https://test.pypi.org/legacy/" },
]
